Fixes #5142 (#5143)

+++ code/modules/reagents/reagent_containers/food/snacks.dm
@@ -60,7 +60,11 @@
 /obj/item/weapon/reagent_containers/food/snacks/attack_self(mob/user as mob)
 	return
 
-/obj/item/weapon/reagent_containers/food/snacks/afterattack(mob/M as mob, mob/user as mob, def_zone)
+/obj/item/weapon/reagent_containers/food/snacks/afterattack(mob/M as mob, mob/user as mob, proximity)
+
+	if(!proximity)
+		return 0
+
 	if(!reagents.total_volume)
 		user << "<span class='danger'>None of [src] left!</span>"
 		qdel(src)
